real-estate-scraper-backend/requirements.txt
# Web scraping and HTTP requests
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
fake-useragent==1.4.0
requests-html==0.10.0

# Data processing and analysis
pandas==2.1.4
numpy==1.24.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# API framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Task scheduling and background jobs
celery==5.3.4
redis==5.0.1

# Configuration and environment
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# Proxy and networking
requests-ip-rotator==1.0.14
proxy-checker==0.2.5

# Data validation and serialization
marshmallow==3.20.1
jsonschema==4.20.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

